App: Shift pathways to separate file
